[Add][Opt]Add a check update parameter&Export the config from the program (#30)
* fix

* [ImgBot] Optimize images

*Total -- 435.32kb -> 276.76kb (36.42%)

/otherfile/images/config.png -- 20.77kb -> 11.67kb (43.81%)
/otherfile/images/history_index.png -- 414.56kb -> 265.09kb (36.05%)


* [Opt][Incomplete]Optimize history processing

* [FIX][Build APP]Rclone

* [ADD]Support Devices History Display

* [OPT][FIX][BUILDAPP]

* [OPT]Float process

* [FIX]build app

* fix Upload Release Assets

* fix Upload Release Assets

* [FIX] device history

* FIX

* [Fix]Fixes high memory usage

* [Add][Opt]Add a check update parameter&Export the config from the program

---------

+++ modules/download/base.go
@@ -17,10 +17,10 @@ var (
 	Version string
 )
 
-func DownloadStatic(basedirectory string, force bool) error {
+func DownloadStatic(workdirectory string, force bool, checkupdate bool) error {
 	directory := "static"
-	if basedirectory != "" {
-		directory = filepath.Join(basedirectory, "static")
+	if workdirectory != "" {
+		directory = filepath.Join(workdirectory, "static")
 	}
 	if force {
 		//删除
@@ -48,36 +48,43 @@ func DownloadStatic(basedirectory string, force bool) error {
 	checkErr(err)
 	logrus.Info("静态资源已存在，版本号为" + string(forntendVersion))
 
-	resp, err := http.Get("https://mrui-api.hzchu.top/v2/api/checkupdate")
+	// 检查更新
+	if checkupdate {
+		resp, err := http.Get("https://mrui-api.hzchu.top/v2/api/checkupdate")
 
-	if err != nil {
-		logrus.Info("无法获取更新信息，跳过检查")
-		return err
-	}
 
-	defer resp.Body.Close()
-	body, err := io.ReadAll(resp.Body)
-	checkErr(err)
-	var result map[string]interface{}
-	json.Unmarshal(body, &result)
-	front := result["front"].(map[string]interface{})
-	frontversion := front["version"]
-	frontchangelog := front["changelog"]
+		if err != nil {
+			logrus.Info("无法获取更新信息，跳过检查")
+			return err
+		}
 
-	backend := result["backend"].(map[string]interface{})
-	backendversion := backend["version"]
-	backendchangelog := front["changelog"]
 
-	if backendversion != string(Version) {
-		message := fmt.Sprintf("后端程序发现新版本(%v)，请及时更新。更新日志：%v", backendversion, backendchangelog)
-		logrus.Info(message)
-	}
+		defer resp.Body.Close()
+		body, err := io.ReadAll(resp.Body)
+		checkErr(err)
+		var result map[string]interface{}
+		json.Unmarshal(body, &result)
+		front := result["front"].(map[string]interface{})
+		frontversion := front["version"]
+		frontchangelog := front["changelog"]
 
-	if frontversion != string(forntendVersion) {
-		message := fmt.Sprintf("前端文件发现新版本(%v)，在前端页面中进行更新。更新日志：%v", frontversion, frontchangelog)
-		logrus.Info(message)
-		os.RemoveAll(directory)
-		downloadfile(directory)
+		backend := result["backend"].(map[string]interface{})
+		backendversion := backend["version"]
+		backendchangelog := front["changelog"]
+
+		if backendversion != string(Version) {
+			message := fmt.Sprintf("后端程序发现新版本(%v)，请及时更新。更新日志：%v", backendversion, backendchangelog)
+			logrus.Info(message)
+		}
+
+		if frontversion != string(forntendVersion) {
+			message := fmt.Sprintf("前端文件发现新版本(%v)，在前端页面中进行更新。更新日志：%v", frontversion, frontchangelog)
+			logrus.Info(message)
+			os.RemoveAll(directory)
+			downloadfile(directory)
+		}
+	} else {
+		logrus.Info("跳过检查更新")
 	}
 	return nil
 }
